refactor: renames wrapper file
